Replace 'common' submodule with the 'crypto' requirement
Could not test because tests are failing due to a missing file called
'serverkeys.pem'. These tests need to be reviewed since they should
require an external file to be read.

The requirements.txt now includes the 'crypto' repository.
All files have been updated to use the library instead of the 'common'
submodule. The 'decorators' and 'middleware' modules were moved to the
'rest_app' package. It makes more sense for these modules to be
associated with the application than with the library.